third_party/blink/web_tests/fast/dom/root-client-size-iframe-expected.txt

Style recalc/layout when reading clientWidth of documentElement in iframe.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS [object Internals] is defined.
PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS clientWidthBefore is 100
PASS clientWidthAfter is 200
PASS successfullyParsed is true

TEST COMPLETE
